// src/app/models/settings.model.ts
export interface GlobalSettings {
  abs: number;
  drift_assist: number;
  esp: number;
  traction_control: number;
}

export interface SpecificSettings {
  aero_distribution: number;
  arb_front: number;
  arb_rear: number;
  brake_balance: number;
  brake_power: number;
  gearbox: number;
  susp_comp_front: number;
  susp_comp_rear: number;
  susp_geom_camber_front: number;
  susp_geom_camber_rear: number;
  susp_reb_front: number;
  susp_reb_rear: number;
  tire_grip_front: number;
  tire_grip_rear: number;
}